car_warning_system/2_raspberry_pi/requirements_pi.txt
opencv-python==4.8.0
numpy==1.24.3
tensorflow==2.12.0
RPi.GPIO==0.7.1
Pillow==10.0.0
